campus_vehicle_violation/requirements.txt
ultralytics>=8.2.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
tqdm>=4.66.0
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
PyMySQL>=1.1.0
